[gnmi] Move native tests to dedicated suite
